So it turns out I'm a dumbass and was reading in a 32-bit integer when it was actually a 64-bit integer. Reading works now ^^

+++ libtfc/TfcFile.cpp
@@ -171,7 +171,7 @@ void TfcFile::analyze() {
                     this->next(HASH_LEN);
 
                     // get start position
-                    tempEntry.pos = this->readUInt32();
+                    tempEntry.pos = this->readUInt64();
 
                     // skip over tags, we don't care about them at the moment
                     uint32_t blobTagCount = this->readUInt32();
